requirements.txt
torch
numpy
pandas
pyyaml
torchvision